+++ apps/server/.env.example
@@ -40,3 +40,8 @@ password: 32ZFY5CnpA0gOqXTTH6wBb19dImrZZ74
 Internal Database URL: postgres://feedahelp_user:32ZFY5CnpA0gOqXTTH6wBb19dImrZZ74@dpg-ck0qngb6fquc739mumrg-a/feedahelp
 External Database URL: postgres://feedahelp_user:32ZFY5CnpA0gOqXTTH6wBb19dImrZZ74@dpg-ck0qngb6fquc739mumrg-a.frankfurt-postgres.render.com/feedahelp
 PSQL_command: PGPASSWORD=32ZFY5CnpA0gOqXTTH6wBb19dImrZZ74 psql -h dpg-ck0qngb6fquc739mumrg-a.frankfurt-postgres.render.com -U feedahelp_user feedahelp
+
+
+NEXT_PUBLIC_EMAIL_HOST='Gmail'
+NEXT_PUBLIC_EMAIL='feedahelp@gmail.com'
+NEXT_PUBLIC_EMAIL_PASSWORD='etddndmdprdposvo'
